YOLO v6-v11 Training/requirements.txt
PyYAML==6.0
torch==2.0.1
ultralytics==8.0.49


